Refactor, add energy lookup function
